Feat: Add class Main and new configurations of database

+++ src/main/java/org/br/edu/ifsp/reviewcine/repository/FilmeRepository.java
@@ -3,5 +3,5 @@ package org.br.edu.ifsp.reviewcine.repository;
 import org.br.edu.ifsp.reviewcine.model.Filme;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FilmeRepository extends JpaRepository<Filme, Integer> {
+public interface FilmeRepository extends JpaRepository<Filme, Long> {
 }
